bfd9000_web/requirements-dev.txt
# Development-only dependencies (type stubs, linters, etc.)
# Install with: pip install -r requirements-dev.txt
-r requirements.txt
django-types>=0.20
pytest>=8.0,<10.0
pytest-django>=4.10,<5.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0
